+++ R/antigen.garnish_predict.R
@@ -1,5 +1,275 @@
 
 
+## ---- make_BLAST_uuid
+#' Internal function to categorize mutant and wild-type peptides by similarity using `BLAST` to calculate neoepitope amplitude.
+#'
+#' @param dti Data table of nmers.
+#'
+#' @export make_BLAST_uuid
+#' @md
+
+make_BLAST_uuid <- function(dti){
+
+
+if (suppressWarnings(system('which blastp 2> /dev/null', intern = TRUE)) %>%
+          length == 0){
+            message("Skipping BLAST because ncbiblast+ is not in PATH")
+          return(dti)
+
+   }
+
+  dt <- dti[pep_type != "wt" & !is.na(pep_type)]
+
+  # blast first to get pairs for non-mutnfs peptides then run nature paper package
+
+    dt[MHC %like% "HLA", spc := "Hu"] %>%
+      .[MHC %like% "H-2", spc := "Ms"]
+
+  # generate fastas to query
+  parallel::mclapply(dt[, spc %>% unique], function(s){
+
+    dt <- dt[spc == s]
+
+    fa_v <- dt[, .SD %>% unique, .SDcols = c("nmer", "nmer_uuid")] %>%
+             .[order(nmer)] %>% .[, nmer]
+
+    names(fa_v) <- dt[, .SD %>%
+      unique, .SDcols = c("nmer", "nmer_uuid")] %>%
+      .[order(nmer)] %>%
+      .[, nmer_uuid]
+
+    AA <- Biostrings::AAStringSet(fa_v, use.names = TRUE)
+    Biostrings::writeXStringSet(AA, file = paste(s, "_nmer_fasta.fa", sep = ""), format = "fasta")
+
+  })
+
+  # run blastp-short for near matches
+
+  if (file.exists("Ms_nmer_fasta.fa"))
+    system(paste0(
+      "blastp -query Ms_nmer_fasta.fa -task blastp-short -db ~/antigen.garnish/mouse.bdb -out msblastpout.csv -num_threads ", parallel::detectCores(),
+      " -outfmt '10 qseqid sseqid qseq qstart qend sseq sstart send length mismatch pident evalue bitscore'"
+      ))
+
+  if (file.exists("Hu_nmer_fasta.fa"))
+    system(paste0(
+      "blastp -query Hu_nmer_fasta.fa -task blastp-short -db ~/antigen.garnish/human.bdb -out hublastpout.csv -num_threads ", parallel::detectCores(),
+      " -outfmt '10 qseqid sseqid qseq qstart qend sseq sstart send length mismatch pident evalue bitscore'"
+    ))
+
+  # read in CSV
+
+    blastdt <- list.files(pattern = "blastpout\\.csv")
+
+    if (length(blastdt) == 0 ||
+        all(file.info(blastdt)$size == 0)){
+      message("No WT similarity matches found by blast.")
+      return(dti)
+    }
+
+    if (any(file.info(blastdt)$size == 0))
+      blastdt <- blastdt %>% file.info %>%
+        data.table::as.data.table(keep.rownames = TRUE) %>%
+          .[size != 0, rn]
+
+
+    blastdt <- lapply(blastdt, data.table::fread) %>%
+                  data.table::rbindlist %>%
+                    data.table::setnames(names(.),
+                                        c("nmer_uuid",
+                                        "ensembl_prot_id",
+                                        "nmer",
+                                        "q_start",
+                                        "q_stop",
+                                        "WT.peptide",
+                                        "s_start",
+                                        "s_end",
+                                        "overlap_length",
+                                        "mismatch_length",
+                                        "pident",
+                                        "evalue",
+                                        "bitscore"))
+
+    if (nrow(blastdt) == 0){
+      message("No WT similarity matches found by blast.")
+      return(dti)
+    }
+
+  # keep 1 AA substitutions only
+
+  blastdt <- blastdt[, nmer := nmer %>%
+                    stringr::str_replace_all(pattern = "-", replacement = "")] %>%
+             .[, WT.peptide := WT.peptide %>%
+                    stringr::str_replace_all(pattern = "-", replacement = "")] %>%
+                  .[nchar(nmer) == nchar(WT.peptide) & mismatch_length == 1]
+
+  if (nrow(blastdt) == 0){
+    message("No WT similarity matches found by blast.")
+    return(dti)
+  }
+
+  blastdt[, highest := max(bitscore), by = "nmer_uuid"]
+
+  # exclude insertions or deletions
+    blastdt <- blastdt[highest == bitscore] %>%
+                  .[, highest := NULL]
+
+  # keep longest alignment
+    blastdt <- blastdt[, best := max(overlap_length), by = "nmer_uuid"] %>%
+                .[best == overlap_length] %>%
+                .[, best := NULL]
+
+  # dedupe
+    blastdt %>% data.table::setkey(nmer_uuid, pident)
+    blastdt %<>% unique(by = c("nmer_uuid", "pident"))
+
+    blastdt <- blastdt[, .SD %>% unique, .SDcols = c("nmer_uuid", "nmer", "WT.peptide")]
+    blastdt[, blast_uuid := uuid::UUIDgenerate(), by = c("nmer_uuid", "WT.peptide")]
+    dti <- merge(dti,
+                 blastdt[, .SD, .SDcols = c("nmer_uuid", "nmer", "blast_uuid")],
+                 by = c("nmer_uuid", "nmer"), all.x = TRUE)
+
+    # to add WT.peptide back to table need nmer, nmer_i, nmer_l (nchar(nmer)), var_uuid, pep_type
+
+    if (!"var_uuid" %chin% names(dti)) dti[, var_uuid := NA %>% as.character]
+    if (!"effect_type" %chin% names(dti)) dti[, effect_type := NA %>% as.character]
+
+    vdt <- dti[, .SD %>% unique, .SDcols = c("nmer_uuid", "nmer", "nmer_i", "nmer_l", "var_uuid", "sample_id", "effect_type", "MHC")]
+
+    vdt <- merge(vdt, blastdt, by = c("nmer", "nmer_uuid"))
+
+    vdt %>% .[, nmer := NULL] %>% .[, nmer_uuid := NULL]
+
+    vdt <- vdt %>%
+            data.table::setnames("WT.peptide", "nmer") %>%
+              .[, nmer_uuid := uuid::UUIDgenerate(), by = c("nmer", "var_uuid")] %>%
+                  unique %>%
+                    .[, pep_type := "wt"]
+
+    dto <- data.table::rbindlist(list(dti, vdt), fill = TRUE, use.names = TRUE)
+
+  # sanity check to make sure no '-' slipped through from blastp
+
+    dto <- dto[!nmer %like% "-"]
+
+  # run blastp-short for iedb matches
+    message("Running blastp for iedb matches of >= 66.67% homology...")
+
+    if (file.exists("Ms_nmer_fasta.fa") & file.exists("Hu_nmer_fasta.fa"))
+            system("cat Ms_nmer_fasta.fa Hu_nmer_fasta.fa > iedb_query.fa")
+
+    if (list.files(pattern = "nmer_fasta") %>% length == 1)
+        file.copy(list.files(pattern = "nmer_fasta"), to = "iedb_query.fa")
+
+    system(paste0(
+    "blastp -query iedb_query.fa -task blastp-short -db ~/antigen.garnish/iedb.bdb -out iedbout.csv -num_threads ", parallel::detectCores(),
+    " -outfmt '10 qseqid sseqid qseq qstart qend sseq sstart send length mismatch pident evalue bitscore'"
+    ))
+
+    blastdt <- list.files(pattern = "iedbout\\.csv")
+
+    if (length(blastdt) == 0 || file.info(blastdt)$size == 0){
+      message(paste("No IEDB matches found, returning BLAST against reference proteome(s) only...."))
+      return(dto)
+    }
+
+    blastdt <- blastdt %>% data.table::fread %>%
+                    data.table::setnames(names(.),
+                                        c("nmer_uuid",
+                                        "IEDB_anno",
+                                        "nmer",
+                                        "q_start",
+                                        "q_stop",
+                                        "WT.peptide",
+                                        "s_start",
+                                        "s_end",
+                                        "overlap_length",
+                                        "mismatch_length",
+                                        "pident",
+                                        "evalue",
+                                        "bitscore"))
+
+  # keep highest bitscore and 66.67% identical match
+
+    blastdt <- blastdt[, nmer := nmer %>% stringr::str_replace_all(pattern = "-", replacement = "")] %>%
+                  .[, WT.peptide := WT.peptide %>% stringr::str_replace_all(pattern = "-", replacement = "")] %>%
+                  .[nchar(nmer) > 7 & nchar(WT.peptide) > 7 & pident >= 66.6]
+
+    if (nrow(blastdt) == 0){
+      message(paste("No IEDB matches found, returning BLAST against reference proteome(s) only...."))
+      return(dto)
+    }
+
+    blastdt[, highest := max(bitscore), by = "nmer_uuid"]
+
+  blastdt <- blastdt[highest == bitscore] %>%
+              .[!(nmer %like% "-")] %>%
+                .[, highest := NULL]
+
+  # keep longest alignment
+
+   blastdt <- blastdt[, best := max(overlap_length), by = "nmer_uuid"] %>%
+              .[best == overlap_length] %>%
+                .[, best := NULL]
+
+  # dedupe
+    blastdt %>% data.table::setkey(nmer_uuid, pident)
+    blastdt %<>% unique(by = c("nmer_uuid", "pident"))
+
+  # get full IEDB ref here
+
+    fa <- Biostrings::readAAStringSet("~/antigen.garnish/iedb.fasta")
+    f <- fa %>% data.table::as.data.table %>% .[, x]
+    names(f) <- fa@ranges@NAMES
+    blastdt[, IEDB_anno := parallel::mclapply(IEDB_anno, function(i){
+
+      mv <- f[which(stringr::str_detect(pattern = stringr::fixed(i), names(f)))]
+      mv <- mv[which(stringr::str_detect(pattern = stringr::fixed(WT.peptide), mv))]
+      return(paste(names(mv), WT.peptide, sep = "|"))
+
+    }), by = 1:nrow(blastdt)]
+
+    blastdt <- blastdt[, .SD %>% unique, .SDcols = c("nmer_uuid",
+                                                     "nmer",
+                                                     "WT.peptide",
+                                                     "IEDB_anno",
+                                                     "pident")]
+
+    blastdt[, iedb_uuid := uuid::UUIDgenerate(), by = c("nmer_uuid",
+                                                        "WT.peptide",
+                                                        "IEDB_anno")]
+
+    dto <- merge(dto, blastdt[, .SD, .SDcols = c("nmer_uuid",
+                                                 "iedb_uuid",
+                                                 "IEDB_anno",
+                                                 "pident")],
+           by = c("nmer_uuid"),
+           all.x = TRUE)
+
+    # to add WT.peptide back to table need nmer, nmer_i, nmer_l (nchar(nmer)), var_uuid, effect_type
+
+    vdt <- dto[, .SD %>% unique, .SDcols = c("nmer_uuid", "nmer_i", "nmer_l", "var_uuid", "sample_id", "effect_type", "MHC")]
+
+    vdt <- merge(vdt, blastdt, by = "nmer_uuid")
+
+    vdt %>% .[, nmer := NULL] %>% .[, nmer_uuid := NULL]
+
+    vdt <- vdt %>%
+            data.table::setnames("WT.peptide", "nmer") %>%
+              .[, nmer_uuid := uuid::UUIDgenerate(), by = c("nmer", "var_uuid")] %>%
+                  unique %>%
+                    .[, pep_type := "wt"]
+
+    dto <- data.table::rbindlist(list(dto, vdt), fill = TRUE, use.names = TRUE)
+
+  # sanity check to make sure no '-' slipped through from blastp
+
+    dto <- dto[!(nmer %like% "-")]
+
+    return(dto)
+
+}
+
 
 ## ---- make_DAI_uuid
 #' Internal function to pair peptides from missense sites by a common UUID for DAI calculations
@@ -203,6 +473,25 @@ merge_predictions <- function(l, dt){
         dt[!dai_uuid %>% is.na,
         DAI := Consensus_scores[2] /
           Consensus_scores[1], by = dai_uuid]
+
+      if ("blast_uuid" %chin% names(dt)){
+
+         data.table::setkey(dt, pep_type, blast_uuid)
+
+         dt[!blast_uuid %>% is.na,
+          BLAST_A := Consensus_scores[2] /
+            Consensus_scores[1], by = blast_uuid]
+          }
+
+      if ("iedb_uuid" %chin% names(dt)){
+
+          data.table::setkey(dt, pep_type, iedb_uuid)
+
+          dt[!iedb_uuid %>% is.na,
+          IEDB_A := Consensus_scores[2] /
+            Consensus_scores[1], by = iedb_uuid]
+            }
+
       return(dt)
     }
 
@@ -665,6 +954,8 @@ mcMap(function(x, y) (x %>% as.integer):(y %>% as.integer) %>%
 #' @param assemble Logical. Assemble data table?
 #' @param generate Logical. Generate peptides?
 #' @param predict Logical. Predict binding affinities?
+#' @param blast Logical. Run BLASTp to find wild-type peptide and known IEDB matches?
+#' @param fitness Logical. Run model of [Luksza et al. *Nature* 2017](https://www.ncbi.nlm.nih.gov/pubmed/29132144) to predict neoepitope fitness?
 #' @param humandb Character vector. One of "GRCh37" or "GRCh38".
 #' @param mousedb Character vector. One of "GRCm37" or "GRCm38".
 #' @return A data table of binding predictions including:
@@ -686,8 +977,17 @@ mcMap(function(x, y) (x %>% as.integer):(y %>% as.integer) %>%
 #' * **\*_net**: netMHC prediction tool output
 #' * **mhcflurry_\***: mhcflurry_ prediction tool output
 #' * **mhcnuggets_\***: mhcnuggets_ prediction tool output
+#' * **Consensus_scores**: Average value of MHC binding affinity from all prediction tools that contributed output. 95\% confidence intervals are given by **Upper_CI**, **Lower_CI**.
+#' * **DAI**: Differential agretopicty index, see `garnish_summary` for an explanation of DAI.
+#' * **BLAST_A**: Ratio of consensus binding affinity of mutant peptide / closest single AA mismatch from blastp results. Returned only if `blast = TRUE`.
 #'
-#' as well as a transcript description:
+#' fitness model information [Luksza et al. *Nature* 2017](https://www.ncbi.nlm.nih.gov/pubmed/29132144):
+#' * **ResidueChangeClass**: mutant cDNA sequence
+#' * **A**: Component of the fitness model. Differential MHC affinity of mutant and closest wt peptide, similar to `BLAST_A`.
+#' * **R**: TCR recognition probability, determined by comparison to known epitopes in the IEDB.
+#' * **NeoantigenRecognitionPotential**: Product of A and R. The peptide with the highest value per sample is the dominant neoepitope.
+#'
+#' transcript description:
 #' * description
 #' * start_position
 #' * end_position
@@ -776,6 +1076,10 @@ mcMap(function(x, y) (x %>% as.integer):(y %>% as.integer) %>%
 #'      dt <- antigen.garnish::garnish_predictions(path = path) %T>%
 #'      str
 #' }
+#'
+#' @references
+#' Luksza, M, Riaz, N, Makarov, V, Balachandran VP, et al. 2017. A neoepitope fitness model predicts tumour response to checkpoint blockade immunotherapy. Nature. 23;551(7681):512-516
+#'
 #' @export garnish_predictions
 #' @md
 
@@ -784,13 +1088,14 @@ garnish_predictions <- function(dt = NULL,
                                assemble = TRUE,
                                generate = TRUE,
                                predict = TRUE,
+                               blast = fitness,
+                               fitness = TRUE,
                                humandb = "GRCh38",
                                mousedb = "GRCm38"){
 
-  # remove temporary files on exit
   on.exit({
     message("Removing temporary files")
-    list.files(pattern = "(netMHC|mhcflurry|mhcnuggets).*_[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}\\.csv") %>% file.remove
+    list.files(pattern = "_nmer_fasta\\.fa)|iedb_query.fanetMHC|mhcflurry|mhcnuggets).*_[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}\\.csv") %>% file.remove
   })
 
   if (missing(dt) & missing(path)) stop("dt and path are missing.")
@@ -800,6 +1105,10 @@ garnish_predictions <- function(dt = NULL,
     dt <- rio::import(path) %>%
     data.table::as.data.table
 
+  if (!"data.table" %chin% class(dt))
+    stop("Input must be a data table.")
+
+  if (!"MHC" %chin% names(dt)) stop("Input must include MHC alleles, see ?garnish_predictions")
 
   input_type <- vector()
 
@@ -845,6 +1154,9 @@ dt with peptide:
                                  7 13 14
      MHC                         <same as above>
       ")
+
+if (fitness == TRUE & blast == FALSE)
+    message("Setting blast = TRUE")
 
 if (assemble & input_type == "transcript"){
 
@@ -1111,6 +1423,9 @@ if (generate){
     if (input_type == "peptide")
       dt[, dai_uuid := NA %>% as.character]
 
+  if (blast)
+    dt %<>% make_BLAST_uuid
+
 }
 
 if (predict){
@@ -1151,7 +1466,10 @@ if (predict){
       }
 
    }
+   if (fitness)
+     dt %<>% garnish_fitness
    return(dt)
+
 }