+++ NEWS.md
@@ -1,3 +1,10 @@
+# antigen.garnish 0.0.5
+
+* add ncbi-blast functionality to determine neoepitope near matches
+* add ncbi-blast functionality to determine known IEDB matches
+* implement antigen fitness model of Lukza et al. *Nature* 2017
+
+
 # antigen.garnish 0.0.4
 
 * Summary plots